Rename certain package and traits
