+++ src/client.rs
@@ -2,37 +2,6 @@ use raw::Raw;
 use txn::{Oracle, Snapshot, Timestamp, Transaction, Txn};
 use {Key, KeyRange, KvFuture, KvPair, Value};
 
-#[derive(Clone, Debug, Default, Serialize, Deserialize, PartialEq)]
-#[serde(default)]
-#[serde(rename_all = "kebab-case")]
-pub struct Config {
-    pub pd_endpoints: Vec<String>,
-    pub ca_path: Option<String>,
-    pub cert_path: Option<String>,
-    pub key_path: Option<String>,
-}
-
-impl Config {
-    pub fn new<E>(pd_endpoints: E) -> Self
-    where
-        E: IntoIterator<Item = String>,
-    {
-        Config {
-            pd_endpoints: pd_endpoints.into_iter().collect(),
-            ca_path: None,
-            cert_path: None,
-            key_path: None,
-        }
-    }
-
-    pub fn security(mut self, ca_path: String, cert_path: String, key_path: String) -> Self {
-        self.ca_path = Some(ca_path);
-        self.cert_path = Some(cert_path);
-        self.key_path = Some(key_path);
-        self
-    }
-}
-
 impl Into<Config> for String {
     fn into(self) -> Config {
         Config::new(Some(self))
